Add tests for use of val, var, and lazy val
A new spec file, UsingValAndVarsSpec, has been created in the src/test/scala/com/xyzcorp directory. This file contains unit tests written with ScalaTest that verify the expected behavior of val, var, and lazy val in Scala. This includes tests for immutability of val, mutability of var, and deferred resolution of lazy val.